Move client and add test
