+++ src/lib/cherrypickAndCreateTargetPullRequest/autoMergeNowOrLater.ts
@@ -4,7 +4,7 @@ import {
   enablePullRequestAutoMerge,
   isMissingStatusChecksError,
 } from '../github/v4/enablePullRequestAutoMerge';
-import { GithubV4Exception } from '../github/v4/fetchCommits/graphqlClient';
+import { GithubV4Exception } from '../github/v4/client/graphqlClient';
 import { logger } from '../logger';
 import { ora } from '../ora';
