+++ src/lib/github/v4/disablePullRequestAutoMerge.ts
@@ -1,9 +1,6 @@
 import { graphql } from '../../../graphql/generated';
 import { ValidConfigOptions } from '../../../options/options';
-import {
-  getGraphQLClient,
-  GithubV4Exception,
-} from './fetchCommits/graphqlClient';
+import { getGraphQLClient, GithubV4Exception } from './client/graphqlClient';
 import { fetchPullRequestId } from './fetchPullRequestId2';
 
 export async function disablePullRequestAutoMerge(
