+++ src/lib/github/v4/fetchCommits/fetchCommitBySha.ts
@@ -6,7 +6,7 @@ import {
   Commit,
   parseSourceCommit,
 } from '../../../sourceCommit/parseSourceCommit';
-import { GithubV4Exception, getGraphQLClient } from './graphqlClient';
+import { GithubV4Exception, getGraphQLClient } from '../client/graphqlClient';
 
 export async function fetchCommitBySha(options: {
   accessToken: string;
